requirements.txt
pandas
ipython
python-dotenv
litellm[caching]
tqdm
numpy
scikit-learn